+++ src/planners/trajopt/trajopt_utils/src/config.cpp
@@ -8,9 +8,9 @@ namespace util
     // create boost options_description based on variables, parser
     po::options_description od;
     od.add_options()("help,h", "produce help message");
-    for (int i = 0; i < configs.size(); ++i)
+    for (size_t i = 0; i < configs.size(); ++i)
     {
-      for (int j = 0; j < configs[i].params.size(); ++j)
+      for (size_t j = 0; j < configs[i].params.size(); ++j)
       {
         configs[i].params[j]->addToBoost(od);
       }
